Final Commit
With a lot of help from Micah, Paige, and Max we figured out what was going wrong after spending over a hour and 20 min of trying to fix it. The code was working perfectly fine. The problem was it was making it to big and was forcing it out of the screen so that is was not visible. I went into the CSS and I fixed the size and now it looks good maybe a little lopsided but I didn't want to mess to much with the CSS because I didn't write it originally